requirements.txt
pandas>=1.3.2
scipy>=1.5.4
tensorflow==2.7.0
numpy>=1.20.3
biotite==0.31.0
biopython>=1.79
sewar==0.4.4
opencv-python>=4.5.4.60
annoy==1.17.0
matplotlib>=3.4.3